Client/requirements.txt
python-escpos
pillow
flask

# pywin32 - only needed on Windows
# pywin32

# For camera-based QR scanning (optional - only needed for --camera mode)
# opencv-python
# pyzbar
